Put the ChangeCalculator implementation next to TargetRootsCalculator
