+++ src/python/pants/init/engine_initializer.py
@@ -77,12 +77,10 @@ class LegacyGraphScheduler(datatype(['scheduler', 'symbol_table'])):
 
   def new_session(self):
     session = self.scheduler.new_session()
-    scm = get_scm()
-    change_calculator = EngineChangeCalculator(session, self.symbol_table, scm) if scm else None
-    return LegacyGraphSession(session, self.symbol_table, change_calculator)
+    return LegacyGraphSession(session, self.symbol_table)
 
 
-class LegacyGraphSession(datatype(['scheduler_session', 'symbol_table', 'change_calculator'])):
+class LegacyGraphSession(datatype(['scheduler_session', 'symbol_table'])):
   """A thin wrapper around a SchedulerSession configured with @rules for a symbol table."""
 
   def warm_product_graph(self, target_roots):
